refactor(ui): consolidate DeleteMany component
