+++ packages/ui/src/views/HierarchyList/DocumentListSelection/index.tsx
@@ -2,7 +2,7 @@
 
 import React, { Fragment } from 'react'
 
-import { DeleteMany_v4 } from '../../../elements/DeleteMany/index.js'
+import { DeleteMany } from '../../../elements/DeleteMany/index.js'
 import { useDocumentDrawer } from '../../../elements/DocumentDrawer/index.js'
 import { EditMany_v4 } from '../../../elements/EditMany/index.js'
 import { MoveMany } from '../../../elements/Hierarchy/MoveMany/index.js'
@@ -140,7 +140,7 @@ export const DocumentListSelection: React.FC<DocumentListSelectionProps> = ({
             </Fragment>
           ),
         !disableBulkDelete && (
-          <DeleteMany_v4
+          <DeleteMany
             afterDelete={handleActionSuccess}
             key="bulk-delete"
             modalPrefix="hierarchy-list"
